Bug: Delete function was not able to fetch club despite clubs seemingly being loaded in.
Fix: IT WAS A GSI ISSUE!!!!!!! I am so happy that this was the issue because I was losing my mind on why it was able to fetch clubID and match data but not delete matches. I knew that a GSI is kind of like a "lookup path" for my table so I looked into it.

I looked into GSIs here:
https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/GSI.html#GSI.scenario

So from that I looked around because I remember my first implementation I had matches.js and other handlers use byClub where I just partitoned by key only. But then I asked chat if I can optimize it and it gave me byClubV2 becasue it uses partion key clubId AND sort key startedAt. And this totally slipped my mind so I was trying to use byClub instead of byClubV2 which was causing an inconsistancy with lambda resulting in the 502 error. So the main changes are just changing from byClub to byClubV2 to keep consistent with the other handlers (specifically linked correctly to the environment I setup inside of serverless.yml).

Conclusion: Make sure I am using the GSI I created in the environemnt lol because byClub is not active currently which caused the error...